PML-323/fix: remove redundancy of tests
